assets/css: Adjust top & bottom padding header
Adjust the top and bottom padding of the header of the content pages to
avoid that the header occupy a big part of the screen when the scroll is
at the top of the page.

This allows that when the page is loaded, the content is visible and the
title doesn't almost occupy the full page, so it brings the information
to the user without having to scroll down.